Splitting files and removing automatic fallback to UUID identifier. The UUID identifier is now a separate object and the developers are responsible for handling any failures from the default one
